+++ src/lib.rs
@@ -20,3 +20,6 @@ pub mod state;
 pub mod test;
 pub mod timer;
 pub mod util;
+
+#[cfg(feature = "python")]
+pub mod py;