+++ src/ppu.rs
@@ -3,11 +3,10 @@
 use core::fmt;
 use std::{
     borrow::BorrowMut,
-    cell::RefCell,
     cmp::max,
     convert::TryInto,
     fmt::{Display, Formatter},
-    rc::Rc,
+    sync::{Arc, Mutex},
 };
 
 use crate::{
@@ -145,12 +144,50 @@ impl PaletteInfo {
         }
     }
 
+    pub fn from_colors_hex(name: &str, colors_hex: &str) -> Self {
+        let colors = Self::parse_colors_hex(colors_hex);
+        Self::new(name, colors)
+    }
+
+    pub fn parse_colors_hex(colors_hex: &str) -> Palette {
+        let mut colors = [[0u8; RGB_SIZE]; PALETTE_SIZE];
+        for (index, color) in colors_hex.split(',').enumerate() {
+            let color = color.trim();
+            let color = u32::from_str_radix(color, 16).unwrap_or(0);
+            let r = ((color >> 16) & 0xff) as u8;
+            let g = ((color >> 8) & 0xff) as u8;
+            let b = (color & 0xff) as u8;
+            colors[index] = [r, g, b];
+        }
+        colors
+    }
+
     pub fn name(&self) -> &String {
         &self.name
     }
 
+    /// Returns the colors in RGB format.
     pub fn colors(&self) -> &Palette {
         &self.colors
+    }
+
+    /// Returns the colors in hex format, separated by comma.
+    pub fn colors_hex(&self) -> String {
+        let mut buffer = String::new();
+        let mut is_first = true;
+        for color in self.colors.iter() {
+            let r = color[0];
+            let g = color[1];
+            let b = color[2];
+            let color = (r as u32) << 16 | (g as u32) << 8 | b as u32;
+            if is_first {
+                is_first = false;
+            } else {
+                buffer.push(',');
+            }
+            buffer.push_str(format!("{:06x}", color).as_str());
+        }
+        buffer
     }
 }
 
@@ -545,7 +582,7 @@ pub struct Ppu {
     /// The pointer to the parent configuration of the running
     /// Game Boy emulator, that can be used to control the behaviour
     /// of Game Boy emulation.
-    gbc: Rc<RefCell<GameBoyConfig>>,
+    gbc: Arc<Mutex<GameBoyConfig>>,
 }
 
 #[cfg_attr(feature = "wasm", wasm_bindgen)]
@@ -558,7 +595,7 @@ pub enum PpuMode {
 }
 
 impl Ppu {
-    pub fn new(mode: GameBoyMode, gbc: Rc<RefCell<GameBoyConfig>>) -> Self {
+    pub fn new(mode: GameBoyMode, gbc: Arc<Mutex<GameBoyConfig>>) -> Self {
         Self {
             color_buffer: Box::new([0u8; COLOR_BUFFER_SIZE]),
             shade_buffer: Box::new([0u8; COLOR_BUFFER_SIZE]),
@@ -1231,7 +1268,7 @@ impl Ppu {
         self.gb_mode = value;
     }
 
-    pub fn set_gbc(&mut self, value: Rc<RefCell<GameBoyConfig>>) {
+    pub fn set_gbc(&mut self, value: Arc<Mutex<GameBoyConfig>>) {
         self.gbc = value;
     }
 
@@ -2057,7 +2094,7 @@ impl Default for Ppu {
     fn default() -> Self {
         Self::new(
             GameBoyMode::Dmg,
-            Rc::new(RefCell::new(GameBoyConfig::default())),
+            Arc::new(Mutex::new(GameBoyConfig::default())),
         )
     }
 }